+++ src/pathanimations/astar/diagonals/AllowDiagonals.java
@@ -17,8 +17,8 @@ public class AllowDiagonals implements DiagonalsAllowed {
     @Override
     public List<Point> getNeighbours(Point point) {
         List<Point> neighbours = new ArrayList<>();
-        for (int y = point.y() - 1; y <= point.y() + 1; y++) {
-            for (int x = point.x() - 1; x <= point.x() + 1; x++) {
+        for (double y = point.y() - 1; y <= point.y() + 1; y++) {
+            for (double x = point.x() - 1; x <= point.x() + 1; x++) {
                 Point current = new Point(x, y);
                 if (current.equals(point)) { // don't want itself as neighbour
                     continue;
