+++ src/pathanimations/astar/heuristics/Heuristic.java
@@ -3,5 +3,5 @@ package pathanimations.astar.heuristics;
 import graph.Point;
 
 public interface Heuristic {
-    float getFor(Point p1, Point p2);
+    double getFor(Point p1, Point p2);
 }
